requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
yt-dlp==2026.02.21
mutagen==1.47.0
python-multipart==0.0.22
httpx==0.28.1
requests==2.32.3
pydantic-settings==2.10.1
pytest==8.3.5
pytest-asyncio==0.25.3
anyio==4.9.0
